Zend/zend_types.h: move `zend_string` to `zend_string.h`
It is now possible to include only `zend_string.h` without
`zend_types.h`.
