+++ Zend/zend_alloc.h
@@ -21,12 +21,12 @@
 #ifndef ZEND_ALLOC_H
 #define ZEND_ALLOC_H
 
+#include "zend_portability.h"
 #include "zend_result.h"
 
 #include <stdio.h>
 
 #include "../TSRM/TSRM.h"
-#include "zend.h"
 
 #ifndef ZEND_MM_ALIGNMENT
 # error "ZEND_MM_ALIGNMENT was not defined during configure"
