+++ Zend/zend_string.c
@@ -16,6 +16,7 @@
    +----------------------------------------------------------------------+
 */
 
+#include "zend_string.h"
 #include "zend.h"
 #include "zend_globals.h"
 #include "zend_rc_debug.h"
